refactor: 포인트 적립 정보 dto의 패키지 위치를 역할에 맞게 변경
